pyproject.toml
[project]
name = "mlsdm-governed-cognitive-memory"
version = "1.0.0"
dependencies = [
    "numpy>=1.26",
    "pydantic>=2.7",
    "fastapi>=0.111",
    "uvicorn>=0.29",
    "prometheus-client",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-instrumentation-fastapi",
    "pyyaml",
    "tenacity",
]

[tool.mypy]
strict = true

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
addopts = "--cov=src --cov-report=html --cov-fail-under=90"
